Development student

+++ src/pojos/StudentGrade.java
@@ -4,7 +4,7 @@ public class StudentGrade extends SimpleGrade {
 
 	float grade = 0;
 	
-	public StudentGrade(int lectureID, int gradeID, String name, int affect, float average, float grade) {
+	public StudentGrade(int lectureID, int gradeID, String name, float affect, float average, float grade) {
 		super(lectureID, gradeID, name, affect, average);
 		this.grade = grade;
 	}
